backend/requirements.txt
Django
djangorestframework
django-cors-headers
python-decouple
PyJWT[crypto]
supabase
clerk-backend-api
psycopg[binary]
svix